Make SwapRequest and SwapResponse types public
These types are used by NDKSwift and need to be public to avoid linking errors.
